# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flight Booking API",
        "description": "API documentation for Flight Booking System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://backend.test",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/admin/schedules": {
            "post": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Create flight schedule",
                "operationId": "33baf6827af071cb84736dca17b3e34a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateScheduleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Schedule created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/ScheduleResponse"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/schedules/{scheduleId}/phase-out": {
            "post": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Deactivate flight schedule",
                "operationId": "a230a5b1ae8f4911a5586e6ac3ef1afb",
                "parameters": [
                    {
                        "name": "scheduleId",
                        "in": "path",
                        "description": "Schedule ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Schedule phased out successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/schedules/{scheduleId}/reactivate": {
            "post": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Reactivate flight schedule",
                "operationId": "a75bc89d6ba6c2fab77aea1c546abda3",
                "parameters": [
                    {
                        "name": "scheduleId",
                        "in": "path",
                        "description": "Schedule ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Schedule reactivated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register user",
                "operationId": "f9ddb48d31411767b13157de000a7836",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register user successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/UserResponse"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get current user profile",
                "operationId": "4192d4a8a4d046b42855dd395162893f",
                "responses": {
                    "200": {
                        "description": "User profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/UserResponse"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update current user profile",
                "operationId": "d33342ab9cbe07a20a131b3bfe3ad0e6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/UserResponse"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No data provided for update"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CreateScheduleRequest": {
                "required": [
                    "route_id",
                    "flight_number",
                    "departure_time",
                    "days_of_week",
                    "aircraft_id"
                ],
                "properties": {
                    "route_id": {
                        "description": "ID of route",
                        "type": "integer",
                        "example": 1
                    },
                    "flight_number": {
                        "description": "Flight number",
                        "type": "string",
                        "maxLength": 10,
                        "example": "VN123"
                    },
                    "departure_time": {
                        "description": "Departure time (H:i:s)",
                        "type": "string",
                        "format": "time",
                        "example": "08:30:00"
                    },
                    "days_of_week": {
                        "description": "Days of week (1=Monday ... 7=Sunday)",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 1
                        },
                        "example": [
                            1,
                            3,
                            5
                        ]
                    },
                    "aircraft_id": {
                        "description": "Aircraft ID",
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "required": [
                    "email",
                    "password",
                    "first_name",
                    "last_name",
                    "phone_number"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@gmail.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "123456"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "Phuc"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Dinh"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "0987654321"
                    }
                },
                "type": "object"
            },
            "UpdateProfileRequest": {
                "title": "Update Profile Request",
                "properties": {
                    "full_name": {
                        "description": "User full name",
                        "type": "string",
                        "maxLength": 50,
                        "example": "Nguyen Van A"
                    },
                    "phone_number": {
                        "description": "User phone number",
                        "type": "string",
                        "maxLength": 20,
                        "example": "0987654321"
                    }
                },
                "type": "object"
            },
            "ApiResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ScheduleResponse": {
                "title": "Schedule Response",
                "required": [
                    "id",
                    "route_id",
                    "flight_number",
                    "departure_time",
                    "days_of_week",
                    "aircraft_id",
                    "is_active",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "route_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "flight_number": {
                        "type": "string",
                        "example": "VN123"
                    },
                    "departure_time": {
                        "type": "string",
                        "format": "time",
                        "example": "08:30:00"
                    },
                    "days_of_week": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 1
                        },
                        "example": [
                            1,
                            3,
                            5
                        ]
                    },
                    "aircraft_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-03-07T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-03-07T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "UserResponse": {
                "title": "User Response",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "keycloak_id": {
                        "type": "string",
                        "example": "d1a9c9a0-12ab-4c8c-9c1a-123456789abc"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "full_name": {
                        "type": "string",
                        "example": "Nguyen Van A"
                    },
                    "role": {
                        "type": "string",
                        "example": "customer"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "0987654321"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-03-07T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-03-07T10:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Schedule",
            "description": "Schedule"
        },
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "User",
            "description": "User"
        }
    ]
}